refactor: split commands into several modules
