+++ src/commands/index.js
@@ -1,0 +1,4 @@
+export * from './files.js';
+export * from './help.js';
+export * from './nwd.js';
+export * from './os.js';
