+++ src/constants.js
@@ -1,11 +1,6 @@
 export const APP_NAME = 'minishell';
 
-export const CMD_LS = 'ls';
-export const CMD_PWD = 'pwd';
-export const CMD_UP = 'up';
 export const CMD_EXIT = '.exit';
-export const CMD_CD = 'cd';
-export const CMD_ENV = 'env';
 
 export const PWD_TEXT = '\nYou are currently in';
 
@@ -17,13 +12,9 @@ export const RESET = '\x1B[0m';
 
 export const BLACK = '\x1B[0;30m';
 export const RED = '\x1B[0;31m';
-// export const BRED = '\x1B[1;31m';
 export const GREEN = '\x1B[0;32m';
 export const YELLOW = '\x1B[0;33m';
 export const BLUE = '\x1B[0;34m';
 export const PURPLE = '\x1B[0;35m';
 export const CYAN = '\x1B[0;36m';
 export const WHITE = '\x1B[0;37m';
-
-// export const GRAY = '\x1B[0;2;96;96;96m';
-export const GRAY = '\x1B[38;5;8m';
